feat: use clear all button for all filters
